+++ source/tessellation/static_voronoi_mesh.cpp
@@ -436,7 +436,7 @@ void StaticVoronoiMesh::build_v()
 	mesh_vertices.clear();
 	mesh_vertices.resize(static_cast<size_t>(Tri.get_length()));
 	edges.reserve(static_cast<size_t>(static_cast<double>(Tri.get_length())*3.5));
-	int N=Tri.GetOriginalLength();
+	int N=static_cast<int>(Tri.GetOriginalLength());
 	for(int i=0;i<N;++i)
 		mesh_vertices[static_cast<size_t>(i)].reserve(7);
 	int Nfacets=static_cast<int>(Tri.get_num_facet());
@@ -467,13 +467,13 @@ void StaticVoronoiMesh::build_v()
 									  Tri.GetFacetRadius(static_cast<int>(to_check.neighbors[static_cast<size_t>(j)]))))
 					{
 						{
-							if(edge_temp.neighbors.first<Tri.GetOriginalLength())
+						  if(edge_temp.neighbors.first<static_cast<int>(Tri.GetOriginalLength()))
 								mesh_vertices[static_cast<size_t>(edge_temp.neighbors.first)].push_back(static_cast<int>(edges.size()));
 							else
 								if(obc->PointIsReflective(Tri.get_point(
 													static_cast<size_t>(edge_temp.neighbors.first))))
 									edge_temp.neighbors.first = -1;
-							if(edge_temp.neighbors.second<Tri.GetOriginalLength())
+						  if(edge_temp.neighbors.second<static_cast<int>(Tri.GetOriginalLength()))
 								mesh_vertices[static_cast<size_t>(edge_temp.neighbors.second)].push_back(static_cast<int>(edges.size()));
 							else
 								if(obc->PointIsReflective(Tri.get_point(
@@ -830,7 +830,7 @@ vector<int> StaticVoronoiMesh::CellIntersectOuterBoundary(vector<Edge> const&box
 
 bool StaticVoronoiMesh::CloseToBorder(int point,int &border)
 {
-	int olength=Tri.GetOriginalLength();
+  int olength=static_cast<int>(Tri.GetOriginalLength());
 	int n=static_cast<int>(mesh_vertices[static_cast<size_t>(point)].size());
 	for(int i=0;i<n;++i)
 	{
@@ -848,7 +848,7 @@ vector<int> StaticVoronoiMesh::GetBorderingCells(vector<int> const& copied,
 	vector<int> const& totest,int tocheck,vector<int> tempresult,int outer)
 {
 	int border,test;
-	int olength=Tri.GetOriginalLength();
+	int olength=static_cast<int>(Tri.GetOriginalLength());
 	tempresult.push_back(tocheck);
 	sort(tempresult.begin(),tempresult.end());
 	int n=static_cast<int>(mesh_vertices[static_cast<size_t>(tocheck)].size());
@@ -906,7 +906,7 @@ void StaticVoronoiMesh::GetRealNeighbor(vector<int> &result,int point) const
 {
 	result.reserve(7);
 	int n=static_cast<int>(mesh_vertices[static_cast<size_t>(point)].size());
-	int olength=Tri.GetOriginalLength();
+	int olength=static_cast<int>(Tri.GetOriginalLength());
 	for(int i=0;i<n;++i)
 	{
 	  if(edges[static_cast<size_t>(mesh_vertices[static_cast<size_t>(point)][static_cast<size_t>(i)])].neighbors.first==point)
@@ -1032,7 +1032,7 @@ void StaticVoronoiMesh::GetCorners(vector<vector<int> > &copied,
 void StaticVoronoiMesh::GetToTest(vector<vector<int> > &copied,vector<vector<int> > &totest)
 {
   int nsides=static_cast<int>(copied.size());
-	int olength=Tri.GetOriginalLength();
+  int olength=static_cast<int>(Tri.GetOriginalLength());
 	// sort the vectors
 	for(int i=0;i<nsides;++i)
 		sort(copied[static_cast<size_t>(i)].begin(),copied[static_cast<size_t>(i)].end());
